[e2e] move parsing common code
