+++ test/new-e2e/utils/payloadtesthelper/common.go
@@ -1,0 +1,28 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+package payloadtesthelper
+
+type GetPayloadResponse struct {
+	Payloads [][]byte `json:"payloads"`
+}
+
+func areTagsSubsetOfOtherTags(tags, otherTags []string) bool {
+	otherTagsSet := tagsToSet(otherTags)
+	for _, tag := range tags {
+		if _, found := otherTagsSet[tag]; !found {
+			return false
+		}
+	}
+	return true
+}
+
+func tagsToSet(tags []string) map[string]struct{} {
+	tagsSet := map[string]struct{}{}
+	for _, tag := range tags {
+		tagsSet[tag] = struct{}{}
+	}
+	return tagsSet
+}
